chore (refs T22614): move admin Component to client/js.
