+++ client/js/components/admin/StatisticsCharts.js
